Stop ignoring `--optimize` in `roc --optimize main.roc`
This is just inconvenient and does not make our api better.
Leads to users thinking that roc does not optimize well.
This also allows `roc dev --optimize` (which is the exact same as `roc --optimize`).
If a user wants to do that, it should be fine.

This will all eventually be replaced by the larger cli rewrite.