linker/macho: warn when headerpad is too small in host executable
Previously, I wrongly assumed that the Roc compiler controls how
the host is being built, but this is not the case. Therefore,
there is no way to automatically enforcing that the host will have
enough headerpad size. It may be possible to re-use the old idea
of shifting everything in file, however, this will necessitate also
shifting everything in memory since we absolutely have to shift
by page size to guarantee original alignment of sections. This then
becomes difficult as we will not only need update segments/sections
with new offsets and virtual addresses but also every rebase and binding
for the dynamic linker (which effectively means parsing the info and
recreating it). There may be even more avenues where updates are
required. Therefore, for the time being, let's nerf the error into
a warning since the linker is still experimental until we investigate
how difficult shifting actually is. If it turns out to be nigh
impossible, we will convert the warning back into a hard error and
the user will have to recompile the host with increased headerpad size.
Boy, this kinda reminds me of how errors are structured when one is
mixing PIC with non-PIC code ("recompile your code with -fPIC" or
similar).