Initial Fuzzing Spike
This is just trying to get an example of fuzzing working.

Zig internal fuzzing is not ready yet (maybe we will switch for 0.15.0).
So we are instead using AFL.

This just gets really basic fuzzing for the cli running.
It doesn't really check anything useful, but it is a decent example.

Later, I want to add tooling for automatically running fuzzing and managing the corpus.
Not sure exactly how it will all work.

Also renames `command.zig` to `cli.zig` for clarity.

This does not yet work in nix.