[tracy] wait for shutdown
TRACY_NO_EXIT is used to force tracy to wait for the profiler to finish collecting data before shutting down.
Sadly, it doesn't work on macos.

As a workaround, directly call into the tracy profiler on exit.
Request shutdown (which will start flushing data) and wait for shutdown to complete.
This makes the equivalent of TRACY_NO_EXIT the default.

Not, this only works on clean shutdown through `main` or `fatal`.
Any other shutdown will miss this codepath and exit before tracy fully dumps data.