Increase default capacities
Based on a discussion with Richard (quite a while ago at this point),
lets default to over-allocating our array sizes. This is in hopes that the normal flow never reallocates.
We really don't use that much memory in total and this enables the compiler to see solid speed gains.
Until we find a reason not to or get better heuristics, lets use memory to save time.